pyproject.toml
[project]
name = "mirdan"
dynamic = ["version"]
description = "AI Code Quality Orchestrator - Automatically transforms developer prompts into high-quality, structured requests"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Sean Corkum" }
]
keywords = ["mcp", "ai", "code-quality", "prompt-engineering", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "fastmcp>=2.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
ast = [
    "tree-sitter>=0.24",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-javascript>=0.23",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
mirdan = "mirdan.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/mirdan/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/mirdan"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "S",      # flake8-bandit (security)
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific
    "PTH",    # use-pathlib
    "ASYNC",  # async best practices
    "PERF",   # perflint
]
ignore = [
    "RUF012",  # Mutable class attributes - intentional for class constants
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # Allow assert in tests
"src/mirdan/core/gatherers/**" = ["S110"]  # Allow silent exception handling in optional gatherers
"src/mirdan/core/client_registry.py" = ["ASYNC109"]  # timeout parameter is intentional API design
"src/mirdan/core/active_orchestrator.py" = ["ASYNC109"]  # timeout parameter is passthrough API design

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
warn_redundant_casts = true
warn_unused_ignores = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.report]
fail_under = 85

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
